Updated scripts to accept multiple weights in same level.
