+++ Galactic Sports/Assembly-CSharp.csproj
@@ -64,7 +64,7 @@
     <Compile Include="Assets\Scripts\LevelLoader.cs" />
     <Compile Include="Assets\Scripts\Player.cs" />
     <Compile Include="Assets\Scripts\ScoreDisplay.cs" />
-    <Compile Include="Assets\Scripts\Weight.cs" />
+    <Compile Include="Assets\Scripts\SingleWeight.cs" />
     <Compile Include="Assets\Scripts\WeightCollection.cs" />
     <None Include="Assets\Fonts\Dogica\dogica_pixel_license.txt" />
     <None Include="Assets\Fonts\Dogica\dogica_license.txt" />
