+++ Galactic Sports/Assets/Scripts/WeightCollection.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5b7df8a8f7458ff409849927d80fee64
+guid: d50243702b9a27446904794c0fc74511
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
